+++ pkg/cluster/backupstorage/s3.go
@@ -1,26 +1,67 @@
 package backupstorage
 
 import (
+	"io/ioutil"
+	"os"
 	"path"
+	"sync"
 
 	backups3 "github.com/coreos/etcd-operator/pkg/backup/s3"
 	"github.com/coreos/etcd-operator/pkg/backup/s3/s3config"
 	"github.com/coreos/etcd-operator/pkg/spec"
 
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/client-go/kubernetes"
 )
+
+const (
+	awsConfigDirPrefix = "/etc/etcd-operator/aws/"
+)
+
+// Work around https://github.com/aws/aws-sdk-go/issues/1258 .
+// We need to set global ENV when creating S3 client. Thus, we need to lock when doing that.
+var awsEnvMutex sync.Mutex
 
 type s3 struct {
 	s3config.S3Context
 	clusterName  string
 	namespace    string
+	credsDir     string
 	backupPolicy spec.BackupPolicy
 	kubecli      kubernetes.Interface
 	s3cli        *backups3.S3
 }
 
 func NewS3Storage(s3Ctx s3config.S3Context, kubecli kubernetes.Interface, clusterName, ns string, p spec.BackupPolicy) (Storage, error) {
-	s3cli, err := backups3.New(s3Ctx.S3Bucket, path.Join(ns, clusterName))
+	prefix := path.Join(ns, clusterName)
+	var dir string
+
+	if p.S3 != nil {
+		dir = path.Join(awsConfigDirPrefix, prefix)
+		if err := os.MkdirAll(dir, 0700); err != nil {
+			return nil, err
+		}
+		credsFile, configFile, err := setupAWSConfig(kubecli, ns, p.S3.AWSSecret, dir)
+		if err != nil {
+			return nil, err
+		}
+		awsEnvMutex.Lock()
+		defer awsEnvMutex.Unlock()
+
+		err = os.Setenv("AWS_SHARED_CREDENTIALS_FILE", credsFile)
+		if err != nil {
+			return nil, err
+		}
+		defer os.Unsetenv("AWS_SHARED_CREDENTIALS_FILE")
+
+		err = os.Setenv("AWS_CONFIG_FILE", configFile)
+		if err != nil {
+			return nil, err
+		}
+		defer os.Unsetenv("AWS_CONFIG_FILE")
+	}
+
+	s3cli, err := backups3.New(s3Ctx.S3Bucket, prefix)
 	if err != nil {
 		return nil, err
 	}
@@ -32,6 +73,7 @@ func NewS3Storage(s3Ctx s3config.S3Context, kubecli kubernetes.Interface, cluste
 		backupPolicy: p,
 		namespace:    ns,
 		s3cli:        s3cli,
+		credsDir:     dir,
 	}
 	return s, nil
 }
@@ -65,5 +107,29 @@ func (s *s3) Delete() error {
 			}
 		}
 	}
+	if s.backupPolicy.S3 != nil {
+		return os.RemoveAll(s.credsDir)
+	}
 	return nil
+}
+
+func setupAWSConfig(kubecli kubernetes.Interface, ns, secret, dir string) (string, string, error) {
+	se, err := kubecli.CoreV1().Secrets(ns).Get(secret, metav1.GetOptions{})
+	if err != nil {
+		return "", "", err
+	}
+	creds := se.Data[spec.AWSSecretCredentialsFileName]
+	credsFile := path.Join(dir, "credentials")
+	config := se.Data[spec.AWSSecretConfigFileName]
+	configFile := path.Join(dir, "config")
+
+	err = ioutil.WriteFile(credsFile, creds, 0600)
+	if err != nil {
+		return "", "", err
+	}
+	err = ioutil.WriteFile(configFile, config, 0600)
+	if err != nil {
+		return "", "", err
+	}
+	return credsFile, configFile, nil
 }